% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.getCCC.R
\name{oss.getCCC}
\alias{oss.getCCC}
\title{Determine optimal caret model based on Lin's Concordance Correlation Coefficient}
\usage{
oss.getCCC(Model)
}
\arguments{
\item{Model}{caret model object}
}
\value{
list of three dataframes
Dataframe 1: Fit statistics for best model based on CCC, including best tuning parameters
Dataframe 2: Holds tuning parameters for the model based on CCC
Dataframe 3: Fit statistics for all tuning parameters for the cross validation
}
\description{
This function allows the user to  use  Lin's Concordance
Correlation Coefficient as the metric to select the
optimal model from cross-validation in the caret package.
}
\examples{
#Calculate fit statistics for the best tune model

oss.getCCC(KeeneSOC)


}
