% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.goof.R
\name{oss.goof}
\alias{oss.goof}
\title{Calculate Model Goodness of Fit Metrics}
\usage{
oss.goof(predicted, observed)
}
\arguments{
\item{predicted}{numeric}

\item{observed}{numeric}
}
\value{
data.frame
}
\description{
This function allows the user to calculate r-square
adjusted r-square, root mean square error, mean
absolute error, bias and Lin's concordance
correlation coefficient from observed vs predicted
data from predictive model outputs for continuous
data
}
\examples{
#Calculate fit statistics between 2 numeric vectors
observed<- c(3,7,4,9,6,2)
predicted<-c(2,5,4,7,3,5)
oss.goof(predicted= predicted, observed= observed)


}
