% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.goofCaret.R
\name{oss.goofCaret}
\alias{oss.goofCaret}
\title{Extract Fit Statistics from best fit caret model}
\usage{
oss.goofCaret(ModelName)
}
\arguments{
\item{ModelName}{caret model object}
}
\value{
data.frame
}
\description{
This function allows the user to extract additional
goodness of fit parameters from predictive models
trained in caret for continuous variables; most
importantly Lin's concordance correlation coefficient.
}
\examples{
#Calculate fit statistics for the best tune model
data(KeeneSOC)
oss.goofCaret(KeeneSOC)


}
