% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.textureN.R
\name{oss.texture.n}
\alias{oss.texture.n}
\title{Determine Texture Class}
\usage{
oss.texture.n(sand, silt, clay)
}
\arguments{
\item{sand}{numeric}

\item{silt}{numeric}

\item{clay}{numeric}
}
\value{
numeric
}
\description{
Use sand, silt and clay data to determine the texture
class in the Canadian System of Soil Classification.
Closely resembles oss.texture, but returns an integer
for creating raster
}
\examples{

#Determine texture class for a single observation
oss.texture.n(67,10,23)

#Determine texture class for multiple observations
dat<- data.frame(sand=c(20,40,80), silt= c(15,30,10), clay= c(65,30,10))
mapply(oss.texture.n,sand=dat$sand, silt=dat$silt, clay=dat$clay)

#or return it as a new column in the data frame
dat$class<- mapply(oss.texture.n,sand=dat$sand, silt=dat$silt, clay=dat$clay)

}
