\name{profiles}
\alias{profiles}
\docType{data}
\title{
  Sample soil profile data from the Ottawa Project Area
}
\description{
  Soil profile analytical data for a sample of the Ottawa Study area
}
\usage{data("profiles")}
\format{
  The format is:
    data.frame
}
\details{
  Ontario Ministry of Agriculture, Food and Rural Affairs
}
\source{
  Unpublished data
}
\references{
  Ontario Ministry of Agriculture, Food and Rural Affairs
}
\examples{
  data(profiles)
  ## maybe str(profiles)
}
\keyword{datasets}
