% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.texture.r.R
\name{oss.texture.r}
\alias{oss.texture.r}
\title{Use particle size analysis raster data to create a texture
class raster based on the the Canadian System of Soil Classification.
If sand fraction data is provided, modifiers (coarse, fine, very fine)
will be assigned to the sands, loamy sands and sandy loams.}
\usage{
oss.texture.r(
  sand,
  silt,
  clay,
  vcs = NULL,
  cs = NULL,
  ms = NULL,
  fs = NULL,
  vfs = NULL
)
}
\arguments{
\item{sand}{RasterLayer, RasterStack or RasterBrick}

\item{silt}{RasterLayer, RasterStack or RasterBrick}

\item{clay}{RasterLayer, RasterStack or RasterBrick}

\item{vcs}{RasterLayer, RasterStack or RasterBrick}

\item{cs}{RasterLayer, RasterStack or RasterBrick}

\item{ms}{RasterLayer, RasterStack or RasterBrick}

\item{fs}{RasterLayer, RasterStack or RasterBrick}

\item{vfs}{RasterLayer, RasterStack or RasterBrick}
}
\value{
list
}
\description{
Use particle size analysis raster data to create a texture
class raster based on the the Canadian System of Soil Classification.
If sand fraction data is provided, modifiers (coarse, fine, very fine)
will be assigned to the sands, loamy sands and sandy loams.
}
\examples{
#create sample data which includes all combinations of sand, silt and clay
dat<- data.frame(expand.grid(sand=seq(0,100,1), silt=seq(0,100,1), clay=seq(0,100,1)))
dat$sum<- dat$sand+dat$silt+dat$clay
dat<- dat[dat$sum==100,]

# assign these values to rasters
sand<- raster::raster(nrows=51, ncols=101, vals=dat$sand)
silt<- raster::raster(nrows=51, ncols=101, vals=dat$silt)
clay<- raster::raster(nrows=51, ncols=101, vals=dat$clay)

#Create sand fraction data for testing
vcs<- clay/(clay+silt+sand+clay+clay)*100
cs<- silt/(clay+silt+sand+clay+clay)*100
ms<- sand/(clay+silt+sand+clay+clay)*100
fs<- clay/(clay+silt+sand+clay+clay)*100
vfs<- clay/(clay+silt+sand+clay+clay)*100

#Create a texture class raster without sand fractions
tex<- oss.texture.r(sand,silt,clay)

#Create a texture class raster without sand fractions
tex<- oss.texture.r(sand,silt,clay, vcs, cs, ms, fs, vfs)


}
