% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oss.texture.r.R
\name{oss.texture.r}
\alias{oss.texture.r}
\title{Use particle size analysis raster data to create a texture
class raster based on the the Canadian System of Soil Classification or USDA.
If sand fraction data is provided, modifiers (coarse, fine, very fine)
will be assigned to the sands, loamy sands and sandy loams.}
\usage{
oss.texture.r(
  sand,
  silt,
  clay,
  vcs = NULL,
  cs = NULL,
  ms = NULL,
  fs = NULL,
  vfs = NULL,
  tri = "USDA"
)
}
\arguments{
\item{sand}{RasterLayer, RasterStack or RasterBrick of sand (0.5 - 2 mm) content either in decimal or percentage (e.g. 0.25 or 25)}

\item{silt}{RasterLayer, RasterStack or RasterBrick of silt (0.002 - 0.05 mm) content either in decimal or percentage (e.g. 0.25 or 25)}

\item{clay}{RasterLayer, RasterStack or RasterBrick of clay (<0.002 mm) content either in decimal or percentage (e.g. 0.25 or 25)}

\item{vcs}{RasterLayer, RasterStack or RasterBrick of very coarse sand (1 - 2 mm) content expressed as percentage of the sum of the sand fractions or portion of total sand}

\item{cs}{RasterLayer, RasterStack or RasterBrick of coarse sand (0.5 - 1 mm) content expressed as percentage of the sum of the sand fractions or portion of total sand}

\item{ms}{RasterLayer, RasterStack or RasterBrick of medium sand (0.25 - 0.50 mm) content expressed as percentage of the sum of the sand fractions or portion of total sand}

\item{fs}{RasterLayer, RasterStack or RasterBrick of fine sand (0.10 - 0.25 mm) content expressed as percentage of the sum of the sand fractions or portion of total sand}

\item{vfs}{RasterLayer, RasterStack or RasterBrick of very fine sand (0.05 - 0.10 mm) content expressed as percentage of the sum of the sand fractions or portion of total sand}

\item{tri}{character, current choices are "CSSC" for Canadian System of Soil Classification (default), or "USDA" for United States Department of Agriculture}
}
\value{
When the input data is RasterLayer, returns a list of two objects:
texture_raster
   RasterLayer of soil texture class as per the Canadian System of Soil Classification

legend
   data frame containing the factor levels for the soil texture class raster

When the input data is a RasterStack or a RasterBrick, the function will return a list of lists.
Each outer list item will represent the list of outputs for each layer of the RasterStack or Rasterbrick.
The inner list will contain the 'texture_raster' and 'legend' objects as described above.
}
\description{
Use particle size analysis raster data to create a texture
class raster based on the the Canadian System of Soil Classification or USDA.
If sand fraction data is provided, modifiers (coarse, fine, very fine)
will be assigned to the sands, loamy sands and sandy loams.
}
\examples{
require(sp)
require(raster)
require(rasterVis)
require(RColorBrewer)

# create sample data which includes all combinations of sand, silt and clay
dat<- data.frame(expand.grid(sand=seq(0,100,1), silt=seq(0,100,1), clay=seq(0,100,1)))
dat$sum<- dat$sand+dat$silt+dat$clay
dat<- dat[dat$sum==100,]
dat$x<- dat$sand
dat$y<- dat$clay

coordinates(dat)<- ~x+y
gridded(dat)<- TRUE
sand<- raster(dat[1])
silt<- raster(dat[2])
clay<- raster(dat[3])

# Create sand fraction data for testing
vcs<- clay/(clay+silt+sand+clay+clay)*100
cs<- silt/(clay+silt+sand+clay+clay)*100
ms<- sand/(clay+silt+sand+clay+clay)*100
fs<- clay/(clay+silt+sand+clay+clay)*100
vfs<- clay/(clay+silt+sand+clay+clay)*100

# Create a texture class raster without sand fractions
tex<- oss.texture.r(sand,silt,clay)

# And we can visualize using levelplot
texture.map<- ratify(tex[[1]])
rat<- data.frame(levels(texture.map))
rat[["Texture"]]<- tex[[2]]$Class[match(rat$ID,tex$legend$Code)]
levels(texture.map)<- rat
myTheme<- rasterTheme(region=(colorRampPalette(brewer.pal(12, "Set3"))(13)))
levelplot(texture.map, par.settings=myTheme)

# Create a texture class raster with sand fractions
tex_fractions<- oss.texture.r(sand,silt,clay, vcs, cs, ms, fs, vfs)

}
